app_asistencia/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.0.0
pandas==2.0.3
openpyxl==3.1.2
python-dotenv==1.0.0
Werkzeug==2.3.7
